# docs/swagger.yaml
basePath: /
definitions:
...
  manufaktur.Customer:
    properties:
      customerCode:
        type: string
      dibuatPada:
        type: string
      diubahPada:
        type: string
      id:
        type: integer
      kontak:
        type: string
      nama:
        type: string
    type: object
  manufaktur.DTOBOMSimpan:
    properties:
      idParentMaterial:
        type: integer
      idRawMaterial:
        type: integer
      parameterKuantitasPembentuk:
        type: number
    required:
    - idRawMaterial
    - parameterKuantitasPembentuk
    type: object
  manufaktur.DTOCustomerSimpan:
    properties:
      customerCode:
        type: string
      kontak:
        type: string
      nama:
        type: string
    required:
    - customerCode
    - nama
    type: object
...
  manufaktur.KomposisiMaterialBOM:
    properties:
      dibuatPada:
        type: string
      id:
        type: integer
      idParentMaterial:
        type: integer
      idRawMaterial:
        type: integer
      materialBaku:
        $ref: '#/definitions/manufaktur.Material'
      parameterKuantitasPembentuk:
        type: number
      parentMaterial:
        $ref: '#/definitions/manufaktur.Material'
    type: object
...
  manufaktur.MasterDataSnapshotDto:
    properties:
      jenisDefect:
        items:
          $ref: '#/definitions/manufaktur.JenisDefectSnapshotDto'
        type: array
      kategoriDefect:
        items:
          $ref: '#/definitions/manufaktur.KategoriDefectSnapshotDto'
        type: array
      lineProduksi:
        items:
          $ref: '#/definitions/manufaktur.LineProduksiSnapshotDto'
        type: array
      material:
        items:
          $ref: '#/definitions/manufaktur.MaterialSnapshotDto'
        type: array
      part:
        items:
          $ref: '#/definitions/manufaktur.PartSnapshotDto'
        type: array
      relasiPartDefect:
        items:
          $ref: '#/definitions/manufaktur.RelasiPartDefectSnapshotDto'
        type: array
      slotWaktu:
        items:
          $ref: '#/definitions/manufaktur.SlotWaktuSnapshotDto'
        type: array
      versiMasterData:
        type: string
    type: object
  manufaktur.Material:
    properties:
      beratGSM:
        type: integer
      dibuatPada:
        type: string
      diubahPada:
        type: string
      id:
        type: integer
      idPemasok:
        type: integer
      kodeSKU:
        type: string
      lebarCM:
        type: number
      namaMaterial:
        type: string
      panjangCM:
        type: number
      pemasok:
        $ref: '#/definitions/manufaktur.Pemasok'
      tebalCM:
        type: number
      unitSatuan:
        type: string
    type: object
...
  respon.DetailKesalahan:
    properties:
      field:
        type: string
      pesan:
        type: string
    type: object
  respon.ResponStandar:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      kesalahan:
        items:
          $ref: '#/definitions/respon.DetailKesalahan'
        type: array
      message:
        type: string
      meta: {}
      metadata: {}
      pesan:
        type: string
      status:
        type: string
      success:
        description: Legacy fields untuk retro-kompatibilitas
        type: boolean
      sukses:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk ekosistem manufaktur dan kontrol kualitas PGNServer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PGNServer API
  version: "1.0"
paths:
...
  /api/v1/boms:
    get:
      description: Menampilkan daftar seluruh relasi BOM beserta detail material terkait
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/manufaktur.KomposisiMaterialBOM'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Ambil Semua BOM
      tags:
      - Manufaktur - BOM
    post:
      consumes:
      - application/json
      description: Menyimpan relasi hirarkis BOM baru ke pangkalan data
      parameters:
      - description: Payload Tambah BOM
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/manufaktur.DTOBOMSimpan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.KomposisiMaterialBOM'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Tambah BOM Baru
      tags:
      - Manufaktur - BOM
  /api/v1/boms/{id}:
    delete:
      description: Menghapus data relasi BOM dari sistem berdasarkan ID
      parameters:
      - description: ID BOM
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Hapus BOM
      tags:
      - Manufaktur - BOM
    get:
      description: Mendapatkan data detail relasi BOM berdasarkan ID
      parameters:
      - description: ID BOM
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.KomposisiMaterialBOM'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Detail BOM
      tags:
      - Manufaktur - BOM
    put:
      consumes:
      - application/json
      description: Memperbarui informasi relasi BOM berdasarkan ID
      parameters:
      - description: ID BOM
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Perbarui BOM
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/manufaktur.DTOBOMSimpan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.KomposisiMaterialBOM'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Perbarui BOM
      tags:
      - Manufaktur - BOM
...
  /api/v1/customers:
    get:
      description: Menampilkan daftar seluruh customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/manufaktur.Customer'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Ambil Semua Customer
      tags:
      - Manufaktur - Customer
    post:
      consumes:
      - application/json
      description: Menyimpan data customer baru ke pangkalan data
      parameters:
      - description: Payload Tambah Customer
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/manufaktur.DTOCustomerSimpan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.Customer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Tambah Customer Baru
      tags:
      - Manufaktur - Customer
  /api/v1/customers/{id}:
    delete:
      description: Menghapus data customer dari sistem berdasarkan ID
      parameters:
      - description: ID Customer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Hapus Customer
      tags:
      - Manufaktur - Customer
    get:
      description: Mendapatkan data detail customer berdasarkan ID
      parameters:
      - description: ID Customer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.Customer'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Detail Customer
      tags:
      - Manufaktur - Customer
    put:
      consumes:
      - application/json
      description: Memperbarui informasi customer berdasarkan ID
      parameters:
      - description: ID Customer
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Perbarui Customer
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/manufaktur.DTOCustomerSimpan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.Customer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Perbarui Customer
      tags:
      - Manufaktur - Customer
...
  /api/v1/master-data/snapshot:
    get:
      description: Mengambil snapshot data master lengkap untuk sinkronisasi offline-first
        klien QControl
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/manufaktur.MasterDataSnapshotDto'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Snapshot master data untuk QControl
      tags:
      - Manufaktur - Master Data
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"